requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
groq>=0.5.0
numpy>=1.26.0
